+++ packages/ranger-vite-config/plugin/index.ts
@@ -1,3 +1,4 @@
 export { completePlugin } from './complete'
 export { httpProxy } from './proxy'
+export { reporterPlugin } from './report'
 export { svgBuilder } from './svg'