+++ packages/ranger-vite-config/plugin/report/index.ts
@@ -1,0 +1,1 @@
+export { reporterPlugin } from './report'